% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Mongoose model}
\usage{
model(name, properties, validator)
}
\arguments{
\item{name}{Name of the model}

\item{properties}{List of properties}

\item{validator}{List of validators}
}
\description{
Mongoose model
}
\examples{
# const mongoose = require('mongoose');
library(mongooser)

# mongoose.connect("mongodb://127.0.0.1:27017/test")
mongooser_connect(
  db = "test",
  url = "mongodb://localhost",
  verbose = FALSE,
  options = mongolite::ssl_options()
)

# const Cat = mongoose.model('Cat', { name: String });
Cat <- model(
  "Cat",
  properties = list(
    name = character,
    age = numeric
  ),
  validator = list(
    name = as.character,
    age = as.numeric
  )
)
Cat$drop()

# const kitty = new Cat({ name: 'Zildjian' });
kitty <- Cat$new(
  list(
    name = "Zildjian",
    age = 12
  )
)
# const kitty = new Cat({ name: 'Zildjian' });
fluffy <- Cat$new(
  list(
    name = "fluffy"
  )
)

# kitty.save().then(() => console.log('meow'));
kitty$save()
fluffy$save()

Cat$find()

Cat$drop()

Food <- model(
  "Food",
  properties = list(
    ingredient = character,
    sticker = character,
    to_reheat = logical
  ),
  validator = list(
    ingredient = as.character,
    sticker = as.character,
    to_reheat = as.logical
  )
)
Food$drop()
f_m <- list(
  ingredient = "carrot",
  sticker = "carrot",
  to_reheat = TRUE
)
monday <- Food$new(
  f_m
)
monday$save()
res <- Food$find()[[1]]
class(res$day) == class(f_m$day)


f_t <- list(
  ingredient = c("carrot", "potato"),
  sticker = "carrot",
  to_reheat = 1
)
tuesday <- Food$new(
  f_t
)
tuesday$save()
Food$find()[[2]]$day
class(Food$find()[[2]]$to_reheat)

Food$find(
  list(
    sticker = "carrot"
  )
)
}
