% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoose.R
\name{Mongoose}
\alias{Mongoose}
\title{Mongoose}
\description{
Mongoose

Mongoose
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connect}{\code{Mongoose$connect()}}
\item \href{#method-model}{\code{Mongoose$model()}}
\item \href{#method-clone}{\code{Mongoose$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\if{latex}{\out{\hypertarget{method-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mongoose$connect(
  db = "test",
  url = "mongodb://localhost",
  verbose = FALSE,
  options = mongolite::ssl_options()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model"></a>}}
\if{latex}{\out{\hypertarget{method-model}{}}}
\subsection{Method \code{model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mongoose$model(name, schema_definition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mongoose$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
