% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mongooser_connect}
\alias{mongooser_connect}
\title{Connect Mongoose}
\usage{
mongooser_connect(
  db = "test",
  url = "mongodb://localhost",
  verbose = FALSE,
  options = mongolite::ssl_options()
)
}
\arguments{
\item{db}{name of database}

\item{url}{address of the mongodb server in mongo connection string
\href{https://docs.mongodb.com/manual/reference/connection-string}{URI format}}

\item{verbose}{emit some more output}

\item{options}{additional connection options such as SSL keys/certs.}
}
\description{
Connect Mongoose
}
